requirements.txt
pyairtable>=3.0.0
google-api-python-client==2.118.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
google-generativeai==0.7.2
python-dotenv==1.0.1
tenacity==8.2.3
schedule==1.2.1
